+++ script/model/sklearn_like_model/callback/EarlyStop.py
